reduz a quantidade de consultas
